Add support for re-use untouched data for sync
If the fetched file is not updated in local file system, fetching
it back will only create a hard link to the original file. The
original file will be thus unlocked.

If the file is newly created in local file system, or the file is
a updated file fetched from global file system, the data will be
copied when syncing.

Change-Id: Ifec942eac537b49a4c82d81aba015058d4a6149b
Signed-off-by: Li Xi <lixi@ddn.com>